move PipelineDatabase.importFromIterator into Pipeline.Database
