+++ CGATPipelines/pipeline_annotations.py
@@ -582,7 +582,6 @@ import CGAT.Database as Database
 import CGAT.Biomart as Biomart
 import CGATPipelines.PipelineGeneset as PipelineGeneset
 import CGATPipelines.PipelineGO as PipelineGO
-import CGATPipelines.PipelineDatabase as PipelineDatabase
 import CGATPipelines.PipelineUCSC as PipelineUCSC
 import CGATPipelines.PipelineKEGG as PipelineKEGG
 import CGAT.Intervals as Intervals
@@ -609,7 +608,6 @@ PARAMS.update(dict([
 # Set parameter dictionary in auxilliary modules
 PipelineGeneset.PARAMS = PARAMS
 PipelineGO.PARAMS = PARAMS
-PipelineDatabase.PARAMS = PARAMS
 PipelineUCSC.PARAMS = PARAMS
 
 
@@ -1023,7 +1021,7 @@ def downloadTranscriptInformation(infile, outfile):
 
     data = filter(lambda x: x['ensembl_gene_id'] in gene_ids, data)
 
-    PipelineDatabase.importFromIterator(
+    P.importFromIterator(
         outfile, tablename,
         data,
         columns=columns,
@@ -1077,7 +1075,7 @@ def downloadEntrezToEnsembl(infile, outfile):
         biomart="ensembl",
         dataset=PARAMS["ensembl_biomart_dataset"])
 
-    PipelineDatabase.importFromIterator(
+    P.importFromIterator(
         outfile,
         tablename,
         data,
@@ -1108,7 +1106,7 @@ def downloadTranscriptSynonyms(infile, outfile):
             "ensembl_biomart_dataset"],
         host=PARAMS["ensembl_biomart_host"])
 
-    PipelineDatabase.importFromIterator(
+    P.importFromIterator(
         outfile,
         tablename,
         data,
