+++ doc/Release.rst
@@ -5,6 +5,13 @@ Release notes
 The CGAT pipeline collection is unreleased. Please use
 the latest version from the repository.
 
+0.1
+---
+
+* Re-organize Pipeline.py as a package
+* Move legacy pipelines OPTIC and GPIPE into separate repository
+
+
 Contributions
 =============
